// frontend/src/generated/visual-contract.generated.ts
export const HTML_SANITIZATION = deepFreeze({
  "allowedTags": [
    "p",
    "br",
    "b",
    "i",
    "u",
    "strong",
    "em",
    "mark",
    "s",
    "del",
    "ins",
    "sub",
    "sup",
    "h1",
    "h2",
    "h3",
    "h4",
    "h5",
    "h6",
    "ul",
    "ol",
    "li",
    "a",
    "img",
    "figure",
    "figcaption",
    "video",
    "audio",
    "source",
    "iframe",
    "div",
    "span",
    "section",
    "article",
    "header",
    "footer",
    "nav",
    "aside",
    "table",
    "thead",
    "tbody",
    "tfoot",
    "tr",
    "th",
    "td",
    "caption",
    "form",
    "input",
    "button",
    "select",
    "option",
    "textarea",
    "label",
    "blockquote",
    "pre",
    "code",
    "hr",
    "details",
    "summary"
  ],
  "allowedAttributes": [
    "id",
    "class",
    "title",
    "lang",
    "dir",
    "style",
    "href",
    "target",
    "rel",
    "src",
    "alt",
    "width",
    "height",
    "loading",
    "controls",
    "autoplay",
    "muted",
    "loop",
    "poster",
    "colspan",
    "rowspan",
    "type",
    "name",
    "value",
    "placeholder",
    "disabled",
    "readonly",
    "checked",
    "frameborder",
    "allow",
    "allowfullscreen",
    "referrerpolicy",
    "sandbox",
    "data-*"
  ],
  "forbiddenTags": [
    "script",
    "object",
    "embed",
    "base",
    "meta",
    "link"
  ],
  "forbiddenAttributes": [
    "onerror",
    "onload",
    "onclick",
    "onmouseover",
    "onfocus",
    "onblur"
  ],
  "iframeHosts": [
    "www.youtube.com",
    "player.vimeo.com",
    "www.youtube-nocookie.com"
  ],
  "videoProviders": {
    "youtube": {
      "standardHosts": [
        "youtube.com",
        "www.youtube.com",
        "m.youtube.com",
        "music.youtube.com"
      ],
      "noCookieHosts": [
        "youtube-nocookie.com",
        "www.youtube-nocookie.com"
      ],
      "shortHosts": [
        "youtu.be",
        "www.youtu.be"
      ],
      "outputHost": "www.youtube.com",
      "noCookieOutputHost": "www.youtube-nocookie.com",
      "idPathSegments": [
        "embed",
        "shorts",
        "v",
        "live"
      ]
    },
    "vimeo": {
      "pageHosts": [
        "vimeo.com",
        "www.vimeo.com"
      ],
      "outputHost": "player.vimeo.com"
    }
  },
  "iframeSandbox": "allow-scripts allow-same-origin allow-presentation",
  "inlineStyleProperties": [
    "color",
    "background-color",
    "font-size",
    "font-family",
    "font-weight",
    "font-style",
    "text-decoration",
    "text-align",
    "line-height",
    "text-transform"
  ],
  "inlineStyleValuePatterns": {
    "color": [
      "^#(?:[0-9a-fA-F]{3,8})$",
      "^rgb\\(",
      "^rgba\\(",
      "^hsl\\(",
      "^hsla\\(",
      "^[a-zA-Z]+$"
    ],
    "background-color": [
      "^#(?:[0-9a-fA-F]{3,8})$",
      "^rgb\\(",
      "^rgba\\(",
      "^hsl\\(",
      "^hsla\\(",
      "^[a-zA-Z]+$"
    ],
    "font-size": [
      "^\\d+(?:\\.\\d+)?(?:px|em|rem|%|pt)$"
    ],
    "font-family": [
      "^[\\w\\s,'\"()-]+$"
    ],
    "font-weight": [
      "^(?:normal|bold|bolder|lighter|[1-9]00)$"
    ],
    "font-style": [
      "^(?:normal|italic|oblique)$"
    ],
    "text-decoration": [
      "^(?:none|underline|line-through|overline)(?:\\s+\\w+)*$"
    ],
    "text-align": [
      "^(?:left|right|center|justify)$"
    ],
    "line-height": [
      "^[\\d.]+(?:px|em|rem|%)?$"
    ],
    "text-transform": [
      "^(?:none|uppercase|lowercase|capitalize)$"
    ]
  }
} as const);
export const URL_SANITIZATION = deepFreeze({
  "navigationSchemes": [
    "http",
    "https"
  ],
  "contentSchemes": [
    "http",
    "https",
    "mailto",
    "tel"
  ],
  "mediaSchemes": [
    "http",
    "https",
    "data"
  ],
  "blockedPuckSchemes": [
    "javascript",
    "data",
    "vbscript",
    "file"
  ],
  "stripControlsPattern": "[\\t\\n\\r]"
} as const);